[Feat, Chore] 가중치(난이도) 설정 완료
- 난이도가 높을 수록 배치하기 까다로운 블럭
- Tool을 float -> int 기반 변경할 수 있게 변경
- 모든 블럭들의 가중치 변경 완료

+++ Assets/00.WorkSpace/GIL/Scripts/Shapes/ShapeData.cs
@@ -12,7 +12,7 @@ namespace _00.WorkSpace.GIL.Scripts.Shapes
 
         [Header("Classic Mode")]
         public int scoreForSpawn = 1;
-        public float chanceForSpawn = 1f;
+        public int chanceForSpawn = 1;
 
         private void OnEnable()
         {
